chore(opapi): refactor files
