+++ opapi/src/section-types-generator/index.ts
@@ -1,0 +1,1 @@
+export { generateTypesBySection } from './section-types-generator'
